Store striped array of vertex ID's
emu::parallel::for_each passes a reference to each invocation of
the worker function, but graph::for_each_vertex should pass the
vertex ID for convenience. Previously we would compute the vertex
ID by subtracting the array base pointer. This was done in a wrapper
around the worker function, which steals registers and doesn't look
nice in the profile. Hopefully this is a little cleaner.