Use a worklist to work on edges dynamically
Often we want to parallelize uniformly across all edges.

One way to do this is for_each_out_edge nested within
for_each_vertex. But no pair of policies really gives the
dynamic behavior that we want. We either create too many threads
or stick a single thread with a huge vertex, leading to load imbalance.

The worklist is a linked list of vertices that allows threads
to atomically grab units of work. Thus we can paralleilize across
all the edges of the graph with a single layer of parallelism.

TODO:
- reintroduce the grouped iteration for BFS.
- documentation
- perf testing